requirements.txt
discord.py[voice]
PyNaCl
yt-dlp
python-dotenv